5.1.4 add controller for page
