+++ src/main.js
@@ -1,129 +1,18 @@
 import UserProfile from './components/user-profile.js';
 import Navigation from './components/navigation.js';
 import Sorting from './components/sorting.js';
-import NoData from './components/no-data.js';
-import FilmContainer from './components/film-container.js';
-import FilmCard from './components/film-card.js';
-import ShowMoreButton from './components/show-more-button.js';
-import FilmDetails from './components/film-details.js';
-
+import PageController from './controllers/page.js';
 import {getRandomInteger} from './utils/common.js';
-import {render, remove} from './utils/render.js';
-import {countFilms, getSortedFilms} from './utils/film-utils.js';
+import {render} from './utils/render.js';
+import {countFilms} from './utils/film-utils.js';
 import {generateFilms} from './mock/film.js';
 
 const FILM_COUNT = 18;
-const TOP_FILM_COUNT = 2;
-const COMMENTED_FILM_COUNT = 2;
-const FILM_COUNT_ON_START = 5;
-const FILM_COUNT_BY_BUTTON = 5;
-
-const SortType = {
-  RATING: `rating`,
-  COMMENTS: `comments`,
-};
 
 const films = generateFilms(FILM_COUNT);
 const quantity = countFilms(films);
 const isDataBaseEmpty = films.length === 0 ? true : false;
-let openedFilmDetails = false;
 
-const repeatRender = (container, movies, iterations) => {
-  for (let i = 0; i < iterations; i++) {
-    renderFilm(container, movies[i]);
-  }
-};
-
-const renderFilm = (place, movie) => {
-  const openFilmDetails = (evt) => {
-    evt.preventDefault();
-
-    if (!openedFilmDetails) {
-      bodyElement.appendChild(filmDetailsComponent.getElement());
-      openedFilmDetails = true;
-      document.addEventListener(`keydown`, onEscKeyDown);
-    }
-  };
-
-  const onEscKeyDown = (evt) => {
-    const isEscKey = evt.key === `Escape` || evt.key === `Esc`;
-
-    if (isEscKey) {
-      evt.preventDefault();
-      bodyElement.removeChild(filmDetailsComponent.getElement());
-      openedFilmDetails = false;
-      document.removeEventListener(`keydown`, onEscKeyDown);
-    }
-  };
-
-  const onFilmPosterClick = (evt) => openFilmDetails(evt);
-  const onFilmTitleClick = (evt) => openFilmDetails(evt);
-  const onFilmCommentsClick = (evt) => openFilmDetails(evt);
-
-  const onCloseButtonClick = (evt) => {
-    evt.preventDefault();
-    bodyElement.removeChild(filmDetailsComponent.getElement());
-    openedFilmDetails = false;
-    document.removeEventListener(`keydown`, onEscKeyDown);
-  };
-
-  const filmComponent = new FilmCard(movie);
-  filmComponent.setPosterClickHandler(onFilmPosterClick);
-  filmComponent.setTitleClickHandler(onFilmTitleClick);
-  filmComponent.setCommentsClickHandler(onFilmCommentsClick);
-
-  const filmDetailsComponent = new FilmDetails(movie);
-  filmDetailsComponent.setCloseButtonClickHandler(onCloseButtonClick);
-
-  render(place, filmComponent);
-};
-
-const renderFilmList = (movies) => {
-  if (isDataBaseEmpty) {
-    const noDataComponent = new NoData();
-    render(mainElement, noDataComponent);
-    return;
-  }
-
-  const filmContainerComponent = new FilmContainer();
-  render(mainElement, filmContainerComponent);
-
-  const filmListElement = filmContainerComponent.getElement().querySelector(`.films-list`);
-  const filmListContainerElement = filmListElement.querySelector(`.films-list__container`);
-
-  let showingFilmCount = FILM_COUNT_ON_START;
-  movies.slice(0, showingFilmCount).forEach((movie) => {
-    renderFilm(filmListContainerElement, movie);
-  });
-
-  const showMoreButtonComponent = new ShowMoreButton();
-  render(filmListElement, showMoreButtonComponent);
-
-  // const showMoreButton = filmListElement.querySelector(`.films-list__show-more`);
-  // showMoreButton.addEventListener(`click`, (evt) => {
-  showMoreButtonComponent.setClickHandler((evt) => {
-    evt.preventDefault();
-    const prevFilmCount = showingFilmCount;
-    showingFilmCount += FILM_COUNT_BY_BUTTON;
-
-    movies.slice(prevFilmCount, showingFilmCount).forEach((movie) => {
-      renderFilm(filmListContainerElement, movie);
-    });
-
-    if (showingFilmCount >= movies.length) {
-      remove(showMoreButtonComponent);
-    }
-  });
-
-  const topFilms = getSortedFilms(films, SortType.RATING);
-  const commentedFilms = getSortedFilms(films, SortType.COMMENTS);
-  const topFilmListElement = filmContainerComponent.getElement().querySelector(`.films-list--top .films-list__container`);
-  const commentedFilmListElement = filmContainerComponent.getElement().querySelector(`.films-list--commented .films-list__container`);
-  repeatRender(topFilmListElement, topFilms, TOP_FILM_COUNT);
-  repeatRender(commentedFilmListElement, commentedFilms, COMMENTED_FILM_COUNT);
-};
-
-const bodyElement = document.querySelector(`body`);
 const headerElement = document.querySelector(`.header`);
 const mainElement = document.querySelector(`.main`);
 
@@ -135,7 +24,8 @@ render(headerElement, userProfileComponent);
 render(mainElement, navigationComponent);
 render(mainElement, sortingComponent);
 
-renderFilmList(films);
+const pageController = new PageController(mainElement);
+pageController.render(films);
 
 const footerStatisticsElement = document.querySelector(`.footer__statistics`);
 const filmCountInfoElement = document.createElement(`p`);
